Drop TracyErrorHandler, rename PsrErrorHandler
